Rectify file structure
