+++ src/main/java/RealLive/RealLive/models/Player.java
@@ -1,8 +1,9 @@
-package models;
+package RealLive.RealLive.models;
 
-public class Admin {
+public class Player {
     private String id;
     private String userName;
+    private int points;
 
     public String getId() {
         return id;
@@ -18,5 +19,13 @@ public class Admin {
 
     public void setUserName(String userName) {
         this.userName = userName;
+    }
+
+    public int getPoints() {
+        return points;
+    }
+
+    public void setPoints(int points) {
+        this.points = points;
     }
 }
