+++ src/main/resources/application.properties
@@ -1,1 +1,4 @@
 spring.application.name=RealLive
+spring.datasource.url=${SPRING_DATASOURCE_URL}
+spring.datasource.username=${SPRING_DATASOURCE_USERNAME}
+spring.datasource.password=${SPRING_DATASOURCE_PASSWORD}
